+++ tempest_fastapi_sdk/__init__.py
@@ -153,7 +153,7 @@ from tempest_fastapi_sdk.webpush import (
     WebPushSubscriptionSchema,
 )
 
-__version__: str = "0.17.0"
+__version__: str = "0.18.0"
 
 __all__: list[str] = [
     "CEP",